Add error handling for database initialization script
Previously the 'scripts/initialize-database.sh' would crash if 'jq' and
'docker' weren't accesible on the developers system. The script shouldn't
handle the installation process since this depends on the developers and
their machine. Thus, a warning and early exit has been added to the
script. Additionally, the var directory will now be created if it hasn't
already when executing said script.